+++ resources/views/home.blade.php
@@ -23,11 +23,18 @@
     @csrf
     @foreach($newsfeedList as $each_newsfeed)
 
-    <div class="margin_eachNewsfeed" data-id="{{ $each_newsfeed->id }}">
+    <div class="margin_eachNewsfeed" data-id="{{ $each_newsfeed->newsfeedsid }}">
+        <img class="circle_img_mini" alt="{{$each_newsfeed->name}}" title="{{$each_newsfeed->name}}" onclick="Visit({{$each_newsfeed->u_id}})"
+            @if($each_newsfeed->picture == "")
+                src="/images/nopic.png"
+            @else
+                src="/{{ $each_newsfeed->picture }}"
+            @endif
+        />
         <span class="nwsfd_u_name">{{ $each_newsfeed->name }}</span>
         <span>{{ $each_newsfeed->created_at }} 說：</span>
             @if(isset($User) && $User->id == $each_newsfeed->u_id)
-            <span style="float:right;"><a href="/admin/newsfeed/{{ $each_newsfeed->id }}/edit">修改備忘錄</a></span>
+            <span style="float:right;"><a href="/admin/newsfeed/{{ $each_newsfeed->id }}/edit">修改一句話</a></span>
             @endif
     </div>
     <div class="body_content">{{ $each_newsfeed->content }}</div>
@@ -105,6 +112,15 @@
                 data: formData,
                 success: function(response) {
                     loadComments(newsfeedId);
+                },
+                error: function(xhr, status, error){
+                    // 處理錯誤情況
+                    if (xhr.status === 401) {
+                        // 顯示警告訊息
+                        alert(xhr.responseJSON.message);
+                        // 如果需要，導向登入頁面
+                        window.location.href = "/user/auth/sign-in?redirect=/newsfeed";
+                    }
                 }
             });
         });
@@ -156,20 +172,22 @@
             var newsfeedId = $(this).closest('.nwsfd_goComment').data('id');
             var c_id = $(this).data('c_id');
             var token= $('input[name="_token"]').val();
-            console.log(newsfeedId+' && '+ c_id);
-            $.ajax({
-                url: '/comments/' + c_id,
-                type: 'DELETE',
-                data:{
-                    _token: token,
-                },
-                success: function(response) {
-                    loadComments(newsfeedId);
-                },
-                error: function(xhr, status, error) {
-                    console.error('Error:', error);
-                }
-            });
+            var confirmY= confirm("確定要刪除這則回應嗎？");
+            if(confirmY){
+                    $.ajax({
+                    url: '/comments/' + c_id,
+                    type: 'DELETE',
+                    data:{
+                        _token: token,
+                    },
+                    success: function(response) {
+                        loadComments(newsfeedId);
+                    },
+                    error: function(xhr, status, error) {
+                        console.error('Error:', error);
+                    }
+                });
+            }
         });
     });
 </script>
